AI/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-postgres
pgvector
sqlalchemy
langchain-text-splitters
pydantic
python-multipart
psycopg2-binary
langchain-groq
langchain-qdrant
qdrant-client

